// sourcecode/BTrak.Packages/snova-trading-module/trading-module-components/trading-components/src/lib/billing/models/contract-template.ts
export class ContractTemplateModel {
    contractTemplateName: string;
    contractTemplateFormJson: string;
    contractTemplateId: string;
    termsAndConditionId: string;
    contractTypeId: string;
    contractTypeName: string;
    isArchived: boolean;
    timeStamp: any;   
    fields: any;
    formBgColor: string;
    formKeys: string;
    templateIds : string;
}